Fix a race that can occur while closing a socket that is connecting
Previously, this code used a non-nil socket as a placeholder in
server.AcquireSocket.  This could lead to a race if server.Close is
called while a socket is connecting.  This commit changes the code to
use a nil socket placeholder instead.  If the socket successfully
connects the nil-placeholder is replaced with the valid socket and
removes the nil-placeholder if the socket fails to connect.